# Refactoring insert and search method for improving perfomance