--- a/router.go
+++ b/router.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"errors"
 	"net/http"
-	"path"
-	"strings"
 )
 
 // Router represents the router which handles routing.
@@ -28,7 +26,6 @@ type route struct {
 var (
 	tmpRoute = &route{}
 
-	// NOTE: want to separate this from the error when the parameter is not found.
 	// Error for not found.
 	ErrNotFound = errors.New("no matching route was found")
 	// Error for method not allowed.
@@ -82,8 +79,7 @@ func (r Router) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 			return
 		}
 	}
-	path := cleanPath(req.URL.Path)
-	action, params, err := r.tree.Search(method, path)
+	action, params, err := r.tree.Search(method, req.URL.Path)
 	if err == ErrNotFound {
 		if r.NotFoundHandler == nil {
 			http.NotFoundHandler().ServeHTTP(w, req)
@@ -120,28 +116,4 @@ func methodNotAllowedHandler() http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusMethodNotAllowed)
 	})
-}
-
-// cleanPath returns the canonical path for p, eliminating . and .. elements.
-// This method borrowed from from net/http package.
-// see https://cs.opensource.google/go/go/+/master:src/net/http/server.go;l=2310;bpv=1;bpt=1
-func cleanPath(p string) string {
-	if p == "" {
-		return "/"
-	}
-	if p[0] != '/' {
-		p = "/" + p
-	}
-	np := path.Clean(p)
-	// path.Clean removes trailing slash except for root;
-	// put the trailing slash back if necessary.
-	if p[len(p)-1] == '/' && np != "/" {
-		// Fast path for common case of p being the string we want:
-		if len(p) == len(np)+1 && strings.HasPrefix(p, np) {
-			np = p
-		} else {
-			np += "/"
-		}
-	}
-	return np
 }