# Description
As a title.

# Changes
- Since regular expressions were always used, the process was split between cases where regular expressions were used and cases where they were not used.
- Stopped using strings.FieldFunc.
- Added benchmark test for Insert method

## Before
```sh
go test -bench=. -cpu=1 -benchmem -count=1
goos: darwin
goarch: arm64
pkg: github.com/bmf-san/goblin
BenchmarkStatic1         9711297               108.3 ns/op            32 B/op          1 allocs/op
BenchmarkStatic5         4751539               253.2 ns/op            96 B/op          1 allocs/op
BenchmarkStatic10        2657132               440.0 ns/op           176 B/op          1 allocs/op
BenchmarkWildCard1       3007450               397.6 ns/op           352 B/op          5 allocs/op
BenchmarkWildCard5        905044              1340 ns/op             559 B/op          9 allocs/op
BenchmarkWildCard10       478514              2417 ns/op             805 B/op         14 allocs/op
BenchmarkRegexp1         3007351               399.5 ns/op           352 B/op          5 allocs/op
BenchmarkRegexp5          857139              1329 ns/op             559 B/op          9 allocs/op
BenchmarkRegexp10         492820              2410 ns/op             806 B/op         14 allocs/op
PASS
ok      github.com/bmf-san/goblin       12.384s
```

## After
```sh
go test -bench=. -cpu=1 -benchmem -count=1
goos: darwin
goarch: arm64
pkg: github.com/bmf-san/goblin
BenchmarkSetRoutes1     1000000000               0.0000015 ns/op               0 B/op          0 allocs/op
BenchmarkSetRoutes5     1000000000               0.0000035 ns/op               0 B/op          0 allocs/op
BenchmarkSetRoutes10    1000000000               0.0000025 ns/op               0 B/op          0 allocs/op
BenchmarkStatic1        18489057                63.07 ns/op            0 B/op          0 allocs/op
BenchmarkStatic5         6925203               183.9 ns/op             0 B/op          0 allocs/op
BenchmarkStatic10        3699958               338.2 ns/op             0 B/op          0 allocs/op
BenchmarkWildCard1       4765615               255.0 ns/op           328 B/op          3 allocs/op
BenchmarkWildCard5       1934539               613.0 ns/op           436 B/op          3 allocs/op
BenchmarkWildCard10       854175              1179 ns/op             559 B/op          3 allocs/op
BenchmarkRegexp1         3206086               374.0 ns/op           336 B/op          4 allocs/op
BenchmarkRegexp5          878718              1294 ns/op             460 B/op          8 allocs/op
BenchmarkRegexp10         502275              2312 ns/op             693 B/op         13 allocs/op
PASS
ok      github.com/bmf-san/goblin       12.633s
```

# Impact range
- N/A

# Operational Requirements
N/A

# Related Issue
<!--- Not obligatory, but write any issues if exists it -->

# Supplement
<!--- Not obligatory -->