@bmf-san 2023-01-08
```go
// Delete the / at head of path. ex. /foo/bar → foo/bar

// not to good
strings.Replace(path, "/", "", 1)

// better
if path[:1] == "/" {
　　path = path[1:]
}
```